Initial rest version.
Rest client that can connect to google cloud instance and database.
Upon a GET request, can query the paragraphs table and return rows equal to the keyword.
We probably want to have a `ilike` statement instead of an `=` statement when comparing the keyword column.

We are using mySQL. So connection queries are general sql queries.

TODO would be to create videos and articles tables, along with their routes. Refactor routes into their own
file, and look into combining the reponse of the json into one big thing that can be sent to front end.

Look into better way of storing environment variables.